// opencdx-dashboard/src/api/logistics/model/country.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface Country
 */
export interface Country {
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'iso2'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'iso3'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'continent'?: Continent;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'fips'?: string;
    /**
     * 
     * @type {number}
     * @memberof Country
     */
    'isoNumeric'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof Country
     */
    'languageCodes'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'phoneCode'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'topLevelInternetDomain'?: string;
    /**
     * 
     * @type {Timestamp}
     * @memberof Country
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof Country
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof Country
     */
    'modifier'?: string;
}

export const Continent = {
    Africa: 'AFRICA',
    Antarctica: 'ANTARCTICA',
    Asia: 'ASIA',
    Europe: 'EUROPE',
    NorthAmerica: 'NORTH_AMERICA',
    Oceania: 'OCEANIA',
    SouthAmerica: 'SOUTH_AMERICA',
    Unrecognized: 'UNRECOGNIZED'
} as const;

export type Continent = typeof Continent[keyof typeof Continent];